+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.3.0
+
+- [Video Player] Added support for headers on video load.
+- [Cast & Airplay] Added option to disable casting per player.
+
 ### 0.2.0
 
 - [Cast] Added Cast support.